+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - The `isReady` property is not participating anymore in the AsyncOperation internal state change.
 - A `NSException` is thrown if an AsyncOperation gets started without being ready.
-- Added `StateObservableOperation` to observe easily the most relevant `Operation` status changes (isReady, isExecuting, isFinished, isCancelled).
+- Added `ObservableOperation` to easily observe all KVC compliant Operation properties.
 - Removed `LoggableOperation`.
 
 ### 5.2.0